pyproject.toml
[project]
name = "trainer-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dj-database-url",
    "django>=5.2.11",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "google-genai>=1.63.0",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.1.1",
    "psycopg2-binary",
    "pyjwt",
    "python-dotenv",
    "supabase",
    "ultralytics>=8.4.14",
]
